Fix typo in test fixing build and warning fixes.

+++ leetcode/1404.cpp
@@ -4,7 +4,7 @@
 
 int solve(std::string s)
 {
-    int l = 0, r = s.length() - 1;
+    int l = 0, r = static_cast<int>(s.length()) - 1;
     int ans = 0;
     while (l < r) {
         ans++;
